chore: reorganize python files
Reorganize for clarity and to follow best practice.
Also add isort for sorting imports.
