+++ requirements.txt
@@ -8,6 +8,7 @@ click==8.1.6
 docutils==0.18.1
 idna==3.4
 imagesize==1.4.1
+isort==5.12.0
 Jinja2==3.1.2
 MarkupSafe==2.1.3
 meson==1.2.2
